requirements.txt
streamlit>=1.28.0
streamlit-authenticator>=0.4.0
pyyaml>=6.0
